Updates to err handling
I forgot to check if the type error was still a problem even if it was in the function guess. I had to add more try and except blocks in the while loop because it would read it every now and then but would then break it anyway. The reason I left the try/exc. blocks in the function though is because I would get an initial value error so I used it to create a two line report to try again. I also assigned the recall of the guess func after the comparisons to player_guess so it would understand where to go back to seamlessly; and I added the tries +=1 to outside the while loop because it sometimes wouldn't add it to the counter if you got it right, especially on getting it right the first time, which may just be my text editor glitching but oh well, it works now.