Updated the 'wanna play' input
I updated it to if the input is not either a lower y or lower n, it'll keep asking for a lower y or n. Part of it was to solve when someone inputs like an R or something else; but also to sort of bypass the .lower() method on the inputs, because it wasn't lowering the letter for the initial wanna_play input but would for the wanna_play input in the while loop. So, I just scrapped the .lower() call and used the while loop.