Many updates
One of the main things I updated was the change to the high score. Now, it shows at the beginning, starting out with a 0 as the beginning score, as all games. It needed to check the high_score after the game is played though to make sure it's not 0. If it is, it'll pop it off, otherwise, it'll display the index 0 of high_score. The problem I was running into, without popping the 0, was not being able to print out high_score[1], since there was no [1]. I had to get it to not only use the 0 at the beginning of the game but also be able to replace it with the real high score (since 0 is technically lower for the sort method). The other big thing I updated was the change from randint(0, 20) to (1, 20). A fellow classmate caught that for me, I was thinking it wouldn't include the beginning point but it does, which led to some bugs but I was completely unaware of why it would sometimes choose 0, so big thanks to them! Some of the other changes were, changing the order of the while loop for wanna_play, it was for some reason not letting me play instantly, even though I used 'y', so I put the if statement at the top and it works now; another change was just a space between the * and the input.